load-secrets: Replace load-secrets process with the actual entrypoint
This commit changes the behavior of `load-secrets` instead of letting it
spawning a new process and waiting for it to finish we replace the
load-secrets process with the entrypoint.

Under the hood instead of using `os/exec` package we use the plain
old `exec`[1] syscall.

This is specifically important when running on a custom container image
lambda which has the Runtime Interface Client as part of its entrypoint.
For lambdas to work properly their main process should be the RIC.

This changes drastically the behavior of this wrapper and are not tested
as a lambda layer.

[1]: https://man7.org/linux/man-pages/man3/exec.3.html