load-secrets: pass execution to handler with execve syscall
This commit changes the behavior of `load-secrets` instead of letting it
spawning a new process and waiting for it to finish it replace the
running process with the entrypoint.

Under the hood instead of using `os/exec` package we use the plain
old `exec`[1] syscall.

This is specifically important when running on a custom container image
lambda which has the Runtime Interface Client as part of its entrypoint.
For lambdas to work properly their main process should be the RIC.