fetch-secrets: Allow argument configuration via ENV variables
When we provision a lambda we might need to get the secrets file from
another directory. For `load-secrets` wrapper this is easily overridable
in the `ENTRYPOINT` of the container image but the same is not true for
the `fetch-secrets` extension which gets invoked automatically.

Hence, we add the support to check for ENV variables in addition to CLI
args.

The precedence is CLI args > ENV variables > defaults.